requirements.txt
click>=8.0.0
loguru>=0.7.0
rich>=13.0.0
osxphotos>=0.72.1
pyyaml>=6.0.0